+++ NEWS.md
@@ -1,7 +1,21 @@
 # text (development versions)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-# text 1.4.9
+# text 1.5.6
+- Improved selection of examples in textExamples()
+- On MacOS setting OMP settings at start-up.
+
+# text 1.5.5
+- added bootstrap_difference and replaced "bootstrap" with "bootstrap_overlap" in the textPredictTest() function.
+
+# text 1.5.4
+- changing `textTrainExamples()` to `textExamples()` and improving the filter_word function. 
+
+# text 1.5.1
+- adding matrix/box legend for `textTrainExamples()`.
+- improving functionality of `textTopics()`.
+
+# text 1.5
 - added `save_output` = "no_plot" in `textTrainRegression()` for  "logistic" and "multinomial" to reduce model size of saved objects.
 
 # text 1.4.8